refactor: Extract notification classes to a separate file
