+++ src/service-worker.ts
@@ -10,35 +10,12 @@
 
 import ExchangeRateApiService from '@common/api/exchange-rate-api.service';
 import { Currency } from '@common/api/exchange-rate-api.types';
+import { ConversionNotification, ExchangeRateNotification } from '@common/notifications';
 import { clientsClaim, setCacheNameDetails } from 'workbox-core';
 import { ExpirationPlugin } from 'workbox-expiration';
 import { createHandlerBoundToURL, precacheAndRoute } from 'workbox-precaching';
 import { registerRoute } from 'workbox-routing';
 import { NetworkFirst, StaleWhileRevalidate } from 'workbox-strategies';
-
-export class ConversionNotification {
-  static readonly type = 'SHOW_CONVERSION_NOTIFICATION';
-
-  constructor(readonly data: { baseCurrency: string; targetCurrency: string; baseCurrencyAmount: number; targetCurrencyAmount: number }) {
-    this.data = data;
-  }
-
-  get type() {
-    return ConversionNotification.type;
-  }
-}
-
-export class ExchangeRateNotification {
-  static readonly type = 'SHOW_EXCHANGE_RATE_NOTIFICATION';
-
-  constructor(readonly data: { baseCurrency: Currency; targetCurrency: Currency }) {
-    this.data = data;
-  }
-
-  get type() {
-    return ExchangeRateNotification.type;
-  }
-}
 
 declare const self: ServiceWorkerGlobalScope;
